+++ client/landing/gutenboarding/section.ts
@@ -5,11 +5,3 @@ export const GUTENBOARDING_SECTION_DEFINITION = {
 	group: 'gutenboarding',
 	enableLoggedOut: true,
 };
-
-export const STEPPER_SECTION_DEFINITION = {
-	name: 'stepper',
-	paths: [ 'stepper' ],
-	module: 'stepper',
-	group: 'stepper',
-	enableLoggedOut: true,
-};
