+++ client/landing/stepper/declarative-flow/internals/steps-repository.tsx
@@ -1,14 +1,17 @@
+import { useI18n } from '@wordpress/react-i18n';
 import type { Step } from './types';
-
 /**
  * The domain picker step
  */
 export const domain: Step = function DomainStep( { navigation } ) {
-	const { goNext } = navigation;
+	const { __ } = useI18n();
+	const { goNext, goBack } = navigation;
+
 	return (
 		<div>
 			<h1>Domains step</h1>
-			<button onClick={ goNext }>Next</button>
+			<button onClick={ goBack }>{ __( 'Previous' ) }</button>
+			<button onClick={ goNext }>{ __( 'Next' ) }</button>
 		</div>
 	);
 };
@@ -17,12 +20,14 @@ export const domain: Step = function DomainStep( { navigation } ) {
  * The design picker step
  */
 export const design: Step = function DesignStep( { navigation } ) {
-	const { goNext } = navigation;
+	const { __ } = useI18n();
+	const { goNext, goBack } = navigation;
 
 	return (
 		<div>
 			<h1>Design step</h1>
-			<button onClick={ goNext }>Next</button>
+			<button onClick={ goBack }>{ __( 'Previous' ) }</button>
+			<button onClick={ goNext }>{ __( 'Next' ) }</button>
 		</div>
 	);
 };
