+++ client/landing/stepper/section.ts
@@ -1,0 +1,7 @@
+export const STEPPER_SECTION_DEFINITION = {
+	name: 'stepper',
+	paths: [ 'stepper' ],
+	module: 'stepper',
+	group: 'stepper',
+	enableLoggedOut: false,
+};
